restapi_rooms/requirements.txt
connexion >= 2.6.0
connexion[swagger-ui] >= 2.6.0
python_dateutil == 2.6.0
setuptools >= 21.0.0
swagger-ui-bundle >= 0.0.2
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

